Separate out KratosUserService and DefaultUserService implementations
