+++ src/main/java/org/radarbase/management/config/ManagementPortalProperties.kt
@@ -16,6 +16,7 @@ data class ManagementPortalProperties @ConstructorBinding constructor(
     val siteSettings: SiteSettings,
 ) {
     data class IdentityServer @ConstructorBinding constructor(
+        val internal: Boolean = true,
         val serverUrl: String,
         val serverAdminUrl: String,
         val adminEmail: String,
