+++ src/main/java/org/radarbase/management/service/SubjectService.kt
@@ -65,8 +65,6 @@ class SubjectService(
     @Autowired private val passwordService: PasswordService,
     @Autowired private val authorityRepository: AuthorityRepository,
     @Autowired private val authService: AuthService,
-    @Autowired private val userService: UserService,
-    @Autowired private val identityService: IdentityService,
 ) {
     /**
      * Create a new subject.
@@ -115,15 +113,7 @@ class SubjectService(
         sourceRepository.saveAll(subject.sources)
 
         val savedSubject = subjectRepository.save(subject)
-        return subjectMapper.subjectToSubjectReducedProjectDTO(savedSubject).also {
-                userService.getUserWithAuthoritiesByLogin(login = subjectDto.login!!)?.let { user ->
-                    try {
-                        identityService.updateAssociatedIdentity(user, savedSubject)
-                    } catch (ex: Exception) {
-                        log.error("Failed to update associated identity for user {}: {}", user.login, ex.message)
-                    }
-                }
-        }
+        return subjectMapper.subjectToSubjectReducedProjectDTO(savedSubject)
     }
 
     suspend fun createSubject(
@@ -382,15 +372,7 @@ class SubjectService(
         }
         subjectRepository.save(subject)
 
-        return sourceMapper.sourceToMinimalSourceDetailsDTO(assignedSource).also {
-            userService.getUserWithAuthoritiesByLogin(login = subject.user?.login!!)?.let { user ->
-                    try {
-                        identityService.updateAssociatedIdentity(user, subject)
-                    } catch (ex: Exception) {
-                        log.error("Failed to update associated identity for user {}: {}", user.login, ex.message)
-                    }
-            }
-        }
+        return sourceMapper.sourceToMinimalSourceDetailsDTO(assignedSource)
     }
 
     /**
